Added to the Mat4 to get it to work (at least translate and rotate) and added some of my PGG Assignment 2 Code (under the folder PGGAssignment2Code) to allow rendering in openGL to help demonstrate and debug the maths. Have now got 9 Sam's (the model of a guy I used in PGG Assignment 2) displayed to the screen performing rotations. Each Sam displayed is its own matrix but they all use the same object to render.

+++ MathsTest/Maths/Vec3.h
@@ -69,7 +69,7 @@ namespace Maths
 		*/
 		float getLength()
 		{
-			return sqrt((x*x) + (y*y) + (z*z));
+			return float(sqrt((x*x) + (y*y) + (z*z)));
 		}
 	};
